backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.6.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
